Reduce code verbosity a bit
Let's add a method to make throw on error simpler
