+++ csharp/PmTask/Clone/AccountCloneHelper.cs
@@ -89,23 +89,16 @@ public class AccountCloneHelper
     private static async Task CloneCustomers(ProjectManagerClient src, ProjectManagerClient dest, AccountMap map)
     {
         // Source
-        var srcCustomers = await src.ProjectCustomer.RetrieveProjectCustomers();
-        if (!srcCustomers.Success)
-        {
-            throw new Exception($"Problem fetching customers from source: {srcCustomers.Error.Message}");
-        }
-
+        var srcCustomers = await src.ProjectCustomer.RetrieveProjectCustomers()
+            .ThrowOnError("Fetching customers from source");
         Console.Write($"Cloning {srcCustomers.Data.Length} customers... ");
 
         // Dest
-        var destCustomers = await dest.ProjectCustomer.RetrieveProjectCustomers();
-        if (!destCustomers.Success)
-        {
-            throw new Exception($"Problem fetching customers from destination: {destCustomers.Error.Message}");
-        }
+        var destCustomers = await dest.ProjectCustomer.RetrieveProjectCustomers()
+            .ThrowOnError("Fetching customers from destination");
 
         // Execute the sync
-        var results = await SyncData<ProjectCustomerDto>(srcCustomers.Data, destCustomers.Data, map,
+        var results = await SyncHelper.SyncData<ProjectCustomerDto>(srcCustomers.Data, destCustomers.Data, map,
             c => c.Name,
             c => c.Id?.ToString() ?? string.Empty,
             async c =>
@@ -114,122 +107,20 @@ public class AccountCloneHelper
                 var created = await dest.ProjectCustomer.CreateProjectCustomer(new ProjectCustomerCreateDto()
                 {
                     Name = c.Name
-                });
-                if (!created.Success || !created.Data.Id.HasValue)
-                {
-                    throw new Exception($"Project customer not created: {created.Error.Message}");
-                }
-
-                return created.Data.Id.Value.ToString();
+                }).ThrowOnError("Creating Project Customer");
+                return created.Data.Id!.Value.ToString();
             },
             (cSrc, cDest) => cSrc.Name == cDest.Name, async (cSrc, cDest) =>
             {
-                if (!cDest.Id.HasValue)
-                {
-                    throw new Exception("Attempted to delete object with null ID");
-                }
-                var updateResult = await dest.ProjectCustomer.UpdateProjectCustomer(cDest.Id.Value, new ProjectCustomerCreateDto() { Name = cSrc.Name });
-                if (!updateResult.Success)
-                {
-                    throw new Exception($"Project customer not updated: {updateResult.Error.Message}");
-                }
+                await dest.ProjectCustomer.UpdateProjectCustomer(cDest.Id!.Value, new ProjectCustomerCreateDto() { Name = cSrc.Name })
+                    .ThrowOnError("Updating Project Customer");
             },
             async (c) =>
             {
-                if (!c.Id.HasValue)
-                {
-                    throw new Exception("Attempted to delete object with null ID");
-                }
-                var deleteResult = await dest.ProjectCustomer.DeleteProjectCustomer(c.Id.Value);
-                if (!deleteResult.Success)
-                {
-                    throw new Exception($"Project customer not deleted: {deleteResult.Error.Message}");
-                }
+                await dest.ProjectCustomer.DeleteProjectCustomer(c.Id!.Value)
+                    .ThrowOnError("Deleting Project Customer");
             });
         Console.WriteLine(results.ToString());
     }
 
-    private class SyncResults
-    {
-        public int Creates { get; set; }
-        public int Updates { get; set; }
-        public int Deletes { get; set; }
-
-        public override string ToString()
-        {
-            if (Creates + Updates + Deletes == 0)
-            {
-                return "No changes.";
-            }
-            var sb = new StringBuilder();
-            if (Creates > 0)
-            {
-                sb.Append($"Created {Creates}, ");
-            }
-            if (Updates > 0)
-            {
-                sb.Append($"Updated {Updates}, ");
-            }
-            if (Deletes > 0)
-            {
-                sb.Append($"Deleted {Deletes}, ");
-            }
-
-            sb.Length -= 2;
-            return sb.ToString();
-        }
-    }
-    
-    private static async Task<SyncResults> SyncData<T>(T[] src, T[] dest, AccountMap map, 
-        Func<T, string> identityFunc, 
-        Func<T, string> primaryKeyFunc,
-        Func<T, Task<string>> createFunc, 
-        Func<T, T, bool> compareFunc,
-        Func<T, T, Task> updateFunc, 
-        Func<T, Task> deleteFunc)
-    {
-        var results = new SyncResults();
-        
-        // Convert our destination list to a dictionary for fast lookup
-        var destMap = dest.ToDictionary(d => identityFunc(d));
-        var destKeyMap = dest.ToDictionary(d => primaryKeyFunc(d));
-        var keysToDelete = dest.Select(d => primaryKeyFunc(d)).ToList();
-        foreach (var item in src)
-        { 
-            var identityString = identityFunc(item);
-            var primaryKeyString = primaryKeyFunc(item);
-            if (destMap.TryGetValue(identityString, out var matchingItem))
-            {
-                // Since this key matches and is valid, don't delete it after sync completes
-                keysToDelete.Remove(primaryKeyFunc(matchingItem));
-                if (!compareFunc(item, matchingItem))
-                {
-                    await updateFunc(item, matchingItem);
-                    results.Updates++;
-                }
-            }
-            else
-            {
-                var newPrimaryKey = await createFunc(item);
-                results.Creates++;
-                map.Items.Add(new AccountMap.AccountMapItem()
-                {
-                    Category = nameof(T), 
-                    Identity = identityString, 
-                    OriginalPrimaryKey = primaryKeyString,
-                    NewPrimaryKey = newPrimaryKey,
-                });
-            }
-        }
-        
-        // Delete other items that shouldn't continue to exist
-        foreach (var key in keysToDelete)
-        {
-            var itemToDelete = destKeyMap[key];
-            await deleteFunc(itemToDelete);
-            results.Deletes++;
-        }
-
-        return results;
-    }
 }
